Finish
This App just Sample to download the font on the iOS Platform.

Just Use the Objective C.

Version: later .iOS SDK 6.0

DESCRIPTION:

Demonstrates how to download fonts on demand on iOS 6 and later.

On iOS 6, we have added the capability for applications to download
fonts on demand. Besides the
fonts installed with iOS 6, applications can install a list of
additional fonts as necessary. This
list of additional fonts can be found in the following article:

========================================================================
===========================
BUILD REQUIREMENTS:

iOS 6.0 SDK or later

========================================================================
===========================
RUNTIME REQUIREMENTS:

iOS 6.0 or later

========================================================================
===========================
CHANGES FROM PREVIOUS VERSIONS:

Version 1.0
- First version.